+++ src/index.js
@@ -2,6 +2,7 @@ const express = require("express");
 require("./db/mongoose");
 const User = require("./models/user");
 const Task = require("./models/task");
+const userRouter = require("./routers/user");
 const { ObjectId } = require("mongodb");
 
 const app = express();
@@ -10,81 +11,8 @@ const port = process.env.PORT || 3000;
 //Parse incomming data to json.
 app.use(express.json());
 
-app.post("/users", async (req, res) => {
-  try {
-    const user = new User(req.body);
-    await user.save();
-    res.status(201).send(user);
-  } catch (error) {
-    res.status(400).send(error);
-  }
-});
-
-app.get("/users", async (req, res) => {
-  try {
-    const users = await User.find({});
-    res.send(users);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-});
-
-app.get("/users/:id", async (req, res) => {
-  try {
-    const user = await User.findById(req.params.id);
-    if (user) {
-      res.send(user);
-    } else {
-      res.status(404).send();
-    }
-  } catch (error) {
-    res.status(500).send(error);
-  }
-});
-
-app.patch("/users/:id", async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ["name", "email", "password", "age"];
-  const isValidOperation = updates.every((update) =>
-    allowedUpdates.includes(update)
-  );
-
-  try {
-    if (!isValidOperation) {
-      return res
-        .status(400)
-        .send({ error: "No properties allowed to update." });
-    }
-
-    const user = await User.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
-
-    if (!user) {
-      return res.status(404).send();
-    }
-
-    return res.send(user);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send();
-  }
-});
-
-app.delete("/users/:id", async (req, res) => {
-  try {
-    const user = await User.findByIdAndDelete(req.params.id);
-
-    if (!user) {
-      return res.status(404).send();
-    }
-
-    return res.send(user);
-  } catch (error) {
-    res.status(500).send();
-  }
-});
+//UserRoutes
+app.use(userRouter);
 
 app.post("/tasks", async (req, res) => {
   try {
