+++ src/routers/user.js
@@ -1,0 +1,81 @@
+const express = require("express");
+const router = new express.Router();
+const User = require("../models/user");
+
+router.post("/users", async (req, res) => {
+  try {
+    const user = new User(req.body);
+    await user.save();
+    res.status(201).send(user);
+  } catch (error) {
+    res.status(400).send(error);
+  }
+});
+
+router.get("/users", async (req, res) => {
+  try {
+    const users = await User.find({});
+    res.send(users);
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
+router.get("/users/:id", async (req, res) => {
+  try {
+    const user = await User.findById(req.params.id);
+    if (user) {
+      res.send(user);
+    } else {
+      res.status(404).send();
+    }
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
+router.patch("/users/:id", async (req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ["name", "email", "password", "age"];
+  const isValidOperation = updates.every((update) =>
+    allowedUpdates.includes(update)
+  );
+
+  try {
+    if (!isValidOperation) {
+      return res
+        .status(400)
+        .send({ error: "No properties allowed to update." });
+    }
+
+    const user = await User.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
+
+    if (!user) {
+      return res.status(404).send();
+    }
+
+    return res.send(user);
+  } catch (error) {
+    console.error(error);
+    res.status(500).send();
+  }
+});
+
+router.delete("/users/:id", async (req, res) => {
+  try {
+    const user = await User.findByIdAndDelete(req.params.id);
+
+    if (!user) {
+      return res.status(404).send();
+    }
+
+    return res.send(user);
+  } catch (error) {
+    res.status(500).send();
+  }
+});
+
+module.exports = router;
